# api-platform/infra/openapi.yaml
...
openapi: "3.0.1"
info:
  title: City of Alpenglow Developer API
  version: "2.1.0"
  description: >-
    Public REST API for the fictional City of Alpenglow, CO: permit records,
    business licenses, and parks & facilities, each served by its own
    microservice. v2 is current; v1 still works but answers with
    Deprecation/Sunset headers pointing at its successor. All endpoints except
    /v2/status and /v1/ping require an x-api-key header; the shared demo key
    is printed at https://api.demos.planetek.org.
  contact:
    name: Planetek
    url: https://demos.planetek.org
...
paths:
...
  /v1/licenses:
    get:
      tags: [licenses]
      summary: List business licenses (deprecated)
      deprecated: true
      description: Bare-array response with deprecation headers. Use GET /v2/licenses.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: category, in: query, required: false, schema: { $ref: "#/components/schemas/LicenseCategory" } }
        - { name: status, in: query, required: false, schema: { $ref: "#/components/schemas/LicenseStatus" } }
      responses:
        "200":
          description: Up to 50 licenses, no pagination.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/License" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${licenses_uri}
...
  /v2/licenses:
    get:
      tags: [licenses]
      summary: List business licenses
      description: Cursor-paginated envelope. Filter with ?category= and ?status=.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: category, in: query, required: false, schema: { $ref: "#/components/schemas/LicenseCategory" } }
        - { name: status, in: query, required: false, schema: { $ref: "#/components/schemas/LicenseStatus" } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
        - { name: nextToken, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: One page of licenses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/License" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${licenses_uri}
  /v2/licenses/{id}:
    get:
      tags: [licenses]
      summary: Get a business license
      description: Responses carry a strong ETag; repeat the call with If-None-Match and an unchanged record answers 304 with no body.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: LIC-0042 }
        - { name: If-None-Match, in: header, required: false, schema: { type: string }, description: "ETag from a previous response; matching answers 304." }
      responses:
        "200":
          description: The license, enveloped.
          headers:
            ETag:
              schema: { type: string }
              description: Strong validator for If-None-Match on later calls.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/License" }
        "304": { $ref: "#/components/responses/NotModified" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${licenses_uri}
  /v2/facilities:
    get:
      tags: [facilities]
      summary: List facilities
      description: Cursor-paginated envelope. Filter with ?kind=. Hours are on the /hours sub-resource.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: kind, in: query, required: false, schema: { $ref: "#/components/schemas/FacilityKind" } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
        - { name: nextToken, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: One page of facilities.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Facility" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${facilities_uri}
  /v2/facilities/{id}:
    get:
      tags: [facilities]
      summary: Get a facility
      description: Responses carry a strong ETag; repeat the call with If-None-Match and an unchanged record answers 304 with no body.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: FAC-009 }
        - { name: If-None-Match, in: header, required: false, schema: { type: string }, description: "ETag from a previous response; matching answers 304." }
      responses:
        "200":
          description: The facility, enveloped.
          headers:
            ETag:
              schema: { type: string }
              description: Strong validator for If-None-Match on later calls.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Facility" }
        "304": { $ref: "#/components/responses/NotModified" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${facilities_uri}
  /v2/facilities/{id}/hours:
    get:
      tags: [facilities]
      summary: Get a facility's hours
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: FAC-009 }
      responses:
        "200":
          description: Weekly hours plus any seasonal note.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/FacilityHours" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${facilities_uri}
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: x-api-key
      in: header
      description: Usage-plan API key. The shared demo-tier key is printed on the docs page; partner-tier keys are issued per integrator.
  responses:
    NotFound:
      description: No record with that id.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotModified:
      description: Not modified — the ETag presented in If-None-Match still matches, so no body is sent.
  schemas:
...
    LicenseCategory:
      type: string
      enum: [food-service, retail, contractor, childcare, liquor, lodging, mobile-vendor]
    LicenseStatus:
      type: string
      enum: [active, expired, suspended]
    FacilityKind:
      type: string
      enum: [park, trail, rec-center, library, pool, sports-field]
...
    License:
      type: object
      properties:
        id: { type: string, example: LIC-0042 }
        businessName: { type: string, example: Powderhound Coffee }
        category: { $ref: "#/components/schemas/LicenseCategory" }
        status: { $ref: "#/components/schemas/LicenseStatus" }
        address: { type: string, example: 210 Cirque St }
        issuedAt: { type: string, example: "2024-05-02" }
        expiresAt: { type: string, example: "2026-05-02" }
    Facility:
      type: object
      properties:
        id: { type: string, example: FAC-009 }
        name: { type: string, example: Cirque Recreation Center }
        kind: { $ref: "#/components/schemas/FacilityKind" }
        address: { type: string, example: 800 Summit Way }
        status: { type: string, example: open }
    FacilityHours:
      type: object
      properties:
        facilityId: { type: string, example: FAC-009 }
        name: { type: string, example: Cirque Recreation Center }
        hours:
          type: object
          description: "mon…sun → opening hours or 'closed'"
          additionalProperties: { type: string }
        seasonalNote: { type: string }
...
    PageMeta:
      type: object
      properties:
        count: { type: integer, description: Items in this page. }
        nextToken: { type: string, description: Opaque cursor for the next page; absent on the last page. }
...
    Error:
      type: object
      properties:
        error: { type: string, example: not_found }
        message: { type: string }